+++ eye_tracker/include/eye_tracker/image/position/ContourPositionEstimator.hpp
@@ -11,7 +11,7 @@ namespace et
     public:
         ContourPositionEstimator(int camera_id);
 
-        bool findPupil(cv::Mat &image, cv::Point2f &pupil_position, float &radius) override;
+        bool findPupil(cv::Mat &image, cv::Point2d &pupil_position, double &radius) override;
 
         bool findGlints(cv::Mat &image, std::vector<cv::Point2f> &glints) override;
 
@@ -21,17 +21,17 @@ namespace et
         std::vector<std::vector<cv::Point>> contours_{};
 
         // Centre of the circle aligned with the hole in the view piece in the image.
-        cv::Point2f *pupil_search_centre_{};
+        cv::Point2d *pupil_search_centre_{};
         // Radius of the circle aligned with the hole in the view piece in the image.
         int *pupil_search_radius_{};
         // Minimal radius of the pupil in pixels.
-        float *min_pupil_radius_{};
+        double *min_pupil_radius_{};
         // Maximal radius of the pupil in pixels.
-        float *max_pupil_radius_{};
+        double *max_pupil_radius_{};
 
-        static inline float euclideanDistance(const cv::Point2f &p, const cv::Point2f &q)
+        static inline double euclideanDistance(const cv::Point2d &p, const cv::Point2d &q)
         {
-            cv::Point2f diff = p - q;
+            cv::Point2d diff = p - q;
             return cv::sqrt(diff.x * diff.x + diff.y * diff.y);
         }
     };
