+++ eye_tracker/include/eye_tracker/optimizers/EyeCentreOptimizer.hpp
@@ -9,20 +9,20 @@ namespace et
     class EyeCentreOptimizer : public cv::DownhillSolver::Function
     {
     public:
-        void setParameters(std::vector<cv::Point3f> front_corners, std::vector<cv::Point3f> back_corners, float alpha,
-                           float beta, float cornea_centre_distance);
+        void setParameters(std::vector<cv::Point3d> front_corners, std::vector<cv::Point3d> back_corners, double alpha,
+                           double beta, double cornea_centre_distance);
 
-        cv::Vec3f getCrossPoint() const;
+        cv::Vec3d getCrossPoint() const;
 
         double calc(const double *x) const override;
     private:
 
         int getDims() const override;
 
-        cv::Vec3f cross_point_{};
-        std::vector<cv::Vec3f> visual_axes_{};
-        std::vector<cv::Vec3f> optical_axes_{};
-        float cornea_centre_distance_{};
+        cv::Vec3d cross_point_{};
+        std::vector<cv::Vec3d> visual_axes_{};
+        std::vector<cv::Vec3d> optical_axes_{};
+        double cornea_centre_distance_{};
     };
 
 } // et
